* Integrated norvig's lispy2 code, got tests passing again and added more tests
