+++ .gitignore
@@ -38,3 +38,5 @@ nosetests.xml
 venv/
 \#*
 .\#*
+
+originals/
