+++ lispy/dialects/haney/special_forms.py
@@ -1,3 +1,9 @@
+from __future__ import print_function
+
 from lispy.dialects.norvig.special_forms import SPECIAL_FORMS as NORVIG_FORMS
 
 SPECIAL_FORMS = NORVIG_FORMS
+SPECIAL_FORMS.update(
+    {
+    }
+)
