+++ lispy/tests/code/basic_syntax.lispy
@@ -1,5 +1,5 @@
 #!/usr/bin/env lispy
 (begin
-        (print (+ 2 2))
-        (print (* 6 8))
-        (print os:environ))
+        (display (+ 2 2))
+        (display (* 6 8))
+        (display os:environ))
