+++ lispy/tests/code/sicp/chapters/1.1.1.lispy
@@ -1,28 +1,28 @@
-(print (+ 137 349))
+(display (+ 137 349))
 ;; 486
 
-(print (- 1000 334))
+(display (- 1000 334))
 ;; 666
 
-(print (* 5 99))
+(display (* 5 99))
 ;; 495
 
-(print (/ 10 5))
+(display (/ 10 5))
 ;; 2
 
-(print (+ 2.7 10))
+(display (+ 2.7 10))
 ;; 12.7
 
-(print (+ 21 35 12 7))
+(display (+ (+ 21 35) (+ 12 7)))
 ;; 75
 
-(print (* 25 4 12))
+(display (* 25 4 12))
 ;; 1200
 
-(print (+ (* 3 5) (- 10 6)))
+(display (+ (* 3 5) (- 10 6)))
 ;; 19
 
-(print
+(display
     (+ (* 3
          (+ (* 2 4)
             (+ 3 5)))
