+++ lispy/tests/code/sicp/chapters/1.1.2.lispy
@@ -1,15 +1,15 @@
 (define size 2)
-(print size)
+(display size)
 ;; 2
 
-(print (* 5 size))
+(display (* 5 size))
 ;; 10
 
 (define pi 3.14159)
 (define radius 10)
-(print (* pi (* radius radius)))
+(display (* pi (* radius radius)))
 ;; 314.159
 
 (define circumference (* 2 pi radius))
-(print circumference)
+(display circumference)
 ;; 62.8318
