+++ lispy/tests/code/sicp/chapters/1.1.3.lispy
@@ -1,3 +1,3 @@
-(print (* (+ 2 (* 4 6))
-   (+ 3 5 7))))
+(display (* (+ 2 (* 4 6))
+   (+ 3 (+ 5 7))))
 ;; 390
