+++ lispy/tests/code/sicp/chapters/1.1.8.lispy
@@ -1,0 +1,15 @@
+(define (sqrt x)
+      (define (square x) (* x x))
+      (define average (lambda (x y) (/ (+ x y) 2)))
+      (define abs (lambda (x) (if (< x 0) (* -1 x) x)))
+      (define (good-enough? guess)
+        (< (abs (- (square guess) x)) 0.001))
+      (define (improve guess)
+        (average guess (/ x guess)))
+      (define (sqrt-iter guess)
+        (if (good-enough? guess)
+            guess
+            (sqrt-iter (improve guess))))
+      (sqrt-iter 1.0))
+(display (sqrt 9))
+;; 3.00009155413
