+++ lispy/tests/code/sicp/chapters/1.2.1.lispy
@@ -1,0 +1,18 @@
+(define (factorial n)
+  (if (= n 1)
+      1
+      (* n (factorial (- n 1)))))
+
+(display (factorial 4))
+;; 24
+
+(define (factorial n)
+  (define iter product counter)
+  (if (> counter n)
+      product
+      (iter (* counter product)
+            (+ counter 1)))
+  (iter 1 1))
+
+(display (factorial 4))
+;; 24
